refactor: move the `Navbar` component file
Moved the `Navbar` component file into the `navbar` directory.

And separates the other components inside the `Navbar` component file into the `components directory inside the `navbar` directory.
