+++ src/components/navbar/Navbar.jsx
@@ -1,24 +1,14 @@
 import { useEffect, useState } from "react";
-import {
-  HiOutlineMenuAlt3,
-  HiOutlineMinus,
-  HiOutlineSearch,
-  HiOutlineX,
-} from "react-icons/hi";
-import { useDispatch, useSelector } from "react-redux";
-import {
-  Link,
-  useLocation,
-  useNavigate,
-  useSearchParams,
-} from "react-router-dom";
+import { HiOutlineMenuAlt3, HiOutlineX } from "react-icons/hi";
+import { Link } from "react-router-dom";
 import classNames from "classnames";
 import { AnimatePresence, motion } from "framer-motion";
 // eslint-disable-next-line import/no-extraneous-dependencies
 import PropTypes from "prop-types";
 
-import { greenpeaceLogoWhite } from "../assets";
-import { updateSearchValue } from "../features/search/searchSlice";
+import { greenpeaceLogoWhite } from "../../assets";
+
+import { NavbarLink, SearchBar } from "./components";
 
 export default function Navbar({ forErrorElement }) {
   const [isUnderLargeSizeWidth, setIsUnderLargeSizeWidth] = useState(
@@ -64,6 +54,7 @@ export default function Navbar({ forErrorElement }) {
       opacity: 1,
     },
   };
+
   const handleResize = () => {
     setIsUnderLargeSizeWidth(window.innerWidth < 1024);
     setNavbarIsOpen(!isUnderLargeSizeWidth);
@@ -155,85 +146,6 @@ export default function Navbar({ forErrorElement }) {
         </AnimatePresence>
       </nav>
     </header>
-  );
-}
-
-function NavbarLink({
-  to,
-  isUnderLargeSizeWidth,
-  setNavbarIsOpen,
-  variants,
-  children,
-}) {
-  return (
-    <motion.li
-      className="group text-2xl font-bold text-white hover:text-white/80 lg:mx-2 lg:text-lg"
-      variants={variants}
-    >
-      <Link
-        to={to}
-        className="text-current"
-        onClick={() => isUnderLargeSizeWidth && setNavbarIsOpen(false)}
-      >
-        {children}
-        <hr className="h-[1px] w-1/2 border-0 bg-current duration-200 group-hover:w-full" />
-      </Link>
-    </motion.li>
-  );
-}
-
-function SearchBar({ isUnderLargeSizeWidth, setNavbarIsOpen }) {
-  const dispatch = useDispatch();
-  const { searchValue } = useSelector((store) => store.search);
-
-  const [searchParams] = useSearchParams();
-  const navigate = useNavigate();
-  const location = useLocation();
-
-  const [value, setValue] = useState(
-    searchValue || searchParams.get("q") || ""
-  );
-
-  const handleChangeSearchValue = (event) => {
-    setValue(event.target.value);
-  };
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-
-    if (value.length === 0) {
-      return;
-    }
-
-    dispatch(updateSearchValue(searchValue));
-    navigate(`/search?q=${value}`);
-
-    if (isUnderLargeSizeWidth) {
-      setNavbarIsOpen(false);
-    }
-  };
-
-  useEffect(() => {
-    if (location.pathname === "/search") {
-      setValue(searchParams.get("q") || "");
-    }
-  }, [searchParams.get("q")]);
-
-  return (
-    <form className="opacity-100" onSubmit={handleSubmit}>
-      <label htmlFor="search-bar" className="flex items-center gap-2">
-        <HiOutlineSearch className="h-6 w-6 text-white" />
-        <HiOutlineMinus className="h-6 w-6 rotate-90 text-white" />
-        <input
-          type="search"
-          id="search-bar"
-          className="form-input border-white bg-transparent text-white placeholder:text-sm placeholder:text-white/60 focus:border-current focus:ring-white/60 md:border-none md:placeholder:text-base md:focus:ring-0"
-          placeholder="Telusuri ..."
-          value={value}
-          onChange={handleChangeSearchValue}
-        />
-      </label>
-    </form>
   );
 }
